Added pydantic based spec documentation for the endpoint
