+++ api/model/patron_auth.py
@@ -1,0 +1,7 @@
+from pydantic import Field
+
+from core.util.flask_util import CustomBaseModel
+
+
+class PatronAuthAccessToken(CustomBaseModel):
+    access_token: str = Field(description="A JWE encrypted access token")
